examples/requirements.txt
python-dotenv==1.0.1
openai==1.61.0
pygame==2.5.2